examples/requirements.txt
accelerate==0.26.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
diffusers==0.25.0
diskcache==5.6.3
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
llama_cpp_python==0.2.33
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pillow==10.2.0
psutil==5.9.7
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scipy==1.11.4
sympy==1.12
tokenizers==0.15.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.1.0
zipp==3.17.0
compel==2.0.2
```
